# contracts/openapi/equipment.json
{
  "components" : {
    "schemas" : {
...
      "GlobalErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "InterventionResponse" : {
        "description" : "Una intervencion registrada en la hoja de vida",
        "properties" : {
          "fechaServicio" : {
            "description" : "Cuando se completo el servicio",
            "example" : "2026-10-04T15:30:00",
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "description" : "Identificador de la intervencion",
            "format" : "uuid",
            "type" : "string"
          },
          "idEquipoCliente" : {
            "description" : "Equipo instalado al que se le hizo",
            "format" : "uuid",
            "type" : "string"
          },
          "idReporteServicio" : {
            "description" : "Reporte de servicio del que sale, para abrirlo desde el historial",
            "format" : "uuid",
            "type" : "string"
          },
          "resultado" : {
            "description" : "Estado en que quedo el equipo",
            "example" : "OPERATIVO",
            "type" : "string"
          },
          "tipoServicio" : {
            "description" : "Tipo de servicio prestado",
            "example" : "PREVENTIVO",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "LifeSheetEmpresa" : {
        "description" : "Viene de la configuracion, no de la base",
        "properties" : {
          "ciudad" : {
            "type" : "string"
          },
          "correo" : {
            "type" : "string"
          },
          "direccion" : {
            "type" : "string"
          },
          "movil" : {
            "type" : "string"
          },
          "nombre" : {
            "type" : "string"
          },
          "telefonos" : {
            "description" : "Telefonos fijos; vacia si no tiene",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "LifeSheetFabricante" : {
        "properties" : {
          "nombre" : {
            "type" : "string"
          },
          "pais" : {
            "description" : "Opcional",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "LifeSheetIdentificacion" : {
        "properties" : {
          "areaServicio" : {
            "type" : "string"
          },
          "ciudadSede" : {
            "type" : "string"
          },
          "cliente" : {
            "type" : "string"
          },
          "codigoInterno" : {
            "description" : "Opcional",
            "type" : "string"
          },
          "correosCliente" : {
            "description" : "Correos vigentes del cliente",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "direccionSede" : {
            "type" : "string"
          },
          "documentoCliente" : {
            "type" : "string"
          },
          "estadoActivo" : {
            "type" : "boolean"
          },
          "fechaCompra" : {
            "format" : "date",
            "type" : "string"
          },
          "idEquipoCliente" : {
            "format" : "uuid",
            "type" : "string"
          },
          "numeroInventario" : {
            "type" : "string"
          },
          "proveedor" : {
            "description" : "Quien vendio esta maquina, opcional",
            "type" : "string"
          },
          "responsables" : {
            "description" : "Encargados del area, o de la sede si el area no tiene; vacia si ninguno",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "sede" : {
            "type" : "string"
          },
          "serie" : {
            "type" : "string"
          },
          "telefonosCliente" : {
            "description" : "Telefonos vigentes del cliente",
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "valorCompra" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "LifeSheetResponse" : {
        "description" : "Hoja de vida de un equipo instalado",
        "properties" : {
          "empresa" : {
            "$ref" : "#/components/schemas/LifeSheetEmpresa",
            "description" : "La empresa que presta el servicio: el membrete del documento"
          },
          "fabricante" : {
            "$ref" : "#/components/schemas/LifeSheetFabricante",
            "description" : "Quien lo fabrico"
          },
          "identificacion" : {
            "$ref" : "#/components/schemas/LifeSheetIdentificacion",
            "description" : "Que unidad es y de quien"
          },
          "servicioTecnico" : {
            "description" : "Historial de intervenciones, de la mas reciente a la mas antigua",
            "items" : {
              "$ref" : "#/components/schemas/InterventionResponse"
            },
            "type" : "array"
          },
          "tecnica" : {
            "$ref" : "#/components/schemas/LifeSheetTecnica",
            "description" : "Que es el equipo: tipo, marca y modelo"
          }
        },
        "type" : "object"
      },
      "LifeSheetTecnica" : {
        "properties" : {
          "definicionTecnica" : {
            "type" : "string"
          },
          "fichaTecnica" : {
            "$ref" : "#/components/schemas/TechnicalSheetResponse",
            "description" : "Del modelo: riesgo, caracteristicas y datos electricos"
          },
          "limpiezaCotidiana" : {
            "description" : "Del tipo de equipo, opcional",
            "type" : "string"
          },
          "marca" : {
            "type" : "string"
          },
          "modelo" : {
            "type" : "string"
          },
          "recomendacionesCuidado" : {
            "type" : "string"
          },
          "registroInvima" : {
            "description" : "Registro INVIMA del modelo, opcional",
            "type" : "string"
          },
          "tecnologiaPredominante" : {
            "type" : "string"
          },
          "tipoEquipo" : {
            "type" : "string"
          },
          "uso" : {
            "description" : "Del tipo de equipo, opcional",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TechnicalSheetResponse" : {
        "properties" : {
          "alimentacion" : {
            "type" : "string"
          },
          "amperaje" : {
            "description" : "En A",
            "type" : "number"
          },
          "caracteristicas" : {
            "type" : "string"
          },
          "frecuencia" : {
            "description" : "En Hz",
            "format" : "int32",
            "type" : "integer"
          },
          "potencia" : {
            "description" : "En W",
            "format" : "int32",
            "type" : "integer"
          },
          "riesgo" : {
            "enum" : [ "I", "IIA", "IIB", "III" ],
            "type" : "string"
          },
          "voltaje" : {
            "description" : "En V",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Alternativa: pegar un token obtenido por otro medio",
        "scheme" : "bearer",
        "type" : "http"
      },
      "keycloak-oauth" : {
        "description" : "Inicia sesion en Keycloak y usa el token en cada peticion",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "http://localhost:8080/realms/malphasos-realm/protocol/openid-connect/auth",
            "scopes" : {
              "email" : "Correo electronico",
              "openid" : "Identificacion del usuario",
              "profile" : "Datos basicos del perfil"
            },
            "tokenUrl" : "http://localhost:8080/realms/malphasos-realm/protocol/openid-connect/token"
          }
        },
        "type" : "oauth2"
      }
    }
  },
  "info" : {
    "description" : "API REST para la gestión de mantenimientos preventivos y clientes",
    "license" : {
      "name" : "GNU GPL v3",
      "url" : "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "title" : "MalphasOS API",
    "version" : "v0.0.1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/v1/api/client-equipments/{id}/life-sheet" : {
      "get" : {
        "description" : "Identificacion, tecnica, fabricante e historial de intervenciones.",
        "operationId" : "getLifeSheet",
        "parameters" : [ {
          "description" : "Identificador del equipo instalado",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LifeSheetResponse"
                }
              }
            },
            "description" : "OK"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Conflict"
          },
          "500" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalErrorResponse"
                }
              }
            },
            "description" : "Internal Server Error"
          }
        },
        "summary" : "Hoja de vida de un equipo",
        "tags" : [ "LifeSheet" ]
      }
    },
...
  },
  "security" : [ {
    "keycloak-oauth" : [ ]
  }, {
    "bearer-jwt" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}